# swagger/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "EQEmu Control Panel API",
    "description": "Used for identifying EQEmu stuff",
    "version": "0.0.1",
    "contact": {
      "name": "support",
      "url": "https://github.com/xackery/eqcp"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/bug": {
      "post": {
        "operationId": "BugCreate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbBugCreateResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbBugCreateRequest"
            }
          }
        ],
        "tags": [
          "BugService"
        ]
      }
    },
    "/v1/bug/search": {
      "get": {
        "operationId": "BugSearch",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbBugSearchResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "orderby",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "orderdesc",
            "in": "query",
            "required": false,
            "type": "boolean",
            "format": "boolean"
          }
        ],
        "tags": [
          "BugService"
        ]
      }
    },
    "/v1/bug/{id}": {
      "get": {
        "operationId": "BugRead",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbBugReadResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "BugService"
        ]
      },
      "delete": {
        "operationId": "BugDelete",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbBugDeleteResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "BugService"
        ]
      },
      "put": {
        "operationId": "BugUpdate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbBugUpdateResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbBugUpdateRequest"
            }
          }
        ],
        "tags": [
          "BugService"
        ]
      },
      "patch": {
        "operationId": "BugPatch",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbBugPatchResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbBugPatchRequest"
            }
          }
        ],
        "tags": [
          "BugService"
        ]
      }
    },
...
  },
  "definitions": {
    "pbBug": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "zone": {
          "type": "string"
        },
        "clientversionid": {
          "type": "string",
          "format": "int64"
        },
        "clientversionname": {
          "type": "string"
        },
        "accountid": {
          "type": "string",
          "format": "int64"
        },
        "characterid": {
          "type": "string",
          "format": "int64"
        },
        "charactername": {
          "type": "string"
        },
        "reporterspoof": {
          "type": "string",
          "format": "int64"
        },
        "categoryid": {
          "type": "string",
          "format": "int64"
        },
        "categoryname": {
          "type": "string"
        },
        "reportername": {
          "type": "string"
        },
        "uipath": {
          "type": "string"
        },
        "posx": {
          "type": "string",
          "format": "int64"
        },
        "posy": {
          "type": "string",
          "format": "int64"
        },
        "posz": {
          "type": "string",
          "format": "int64"
        },
        "heading": {
          "type": "string",
          "format": "int64"
        },
        "timeplayed": {
          "type": "string",
          "format": "int64"
        },
        "targetid": {
          "type": "string",
          "format": "int64"
        },
        "targetname": {
          "type": "string"
        },
        "optionalinfomask": {
          "type": "string",
          "format": "int64"
        },
        "canduplicate": {
          "type": "string",
          "format": "int64"
        },
        "crashbug": {
          "type": "string",
          "format": "int64"
        },
        "targetinfo": {
          "type": "string",
          "format": "int64"
        },
        "characterflags": {
          "type": "string",
          "format": "int64"
        },
        "unknownvalue": {
          "type": "string",
          "format": "int64"
        },
        "bugreport": {
          "type": "string"
        },
        "systeminfo": {
          "type": "string"
        },
        "reportdatetime": {
          "type": "string",
          "format": "int64"
        },
        "bugstatus": {
          "type": "string",
          "format": "int64"
        },
        "lastreview": {
          "type": "string",
          "format": "int64"
        },
        "lastreviewer": {
          "type": "string"
        },
        "reviewernotes": {
          "type": "string"
        }
      }
    },
    "pbBugCreateRequest": {
      "type": "object",
      "properties": {
        "values": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
    "pbBugCreateResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "pbBugDeleteResponse": {
      "type": "object",
      "properties": {
        "rowsaffected": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "pbBugPatchRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "key": {
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      }
    },
    "pbBugPatchResponse": {
      "type": "object",
      "properties": {
        "rowsaffected": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "pbBugReadResponse": {
      "type": "object",
      "properties": {
        "bug": {
          "$ref": "#/definitions/pbBug"
        }
      }
    },
    "pbBugSearchResponse": {
      "type": "object",
      "properties": {
        "Bugs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/pbBug"
          }
        },
        "total": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "pbBugUpdateRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "values": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
    "pbBugUpdateResponse": {
      "type": "object",
      "properties": {
        "rowsaffected": {
          "type": "string",
          "format": "int64"
        }
      }
    },
...
  }
}